cli: Remove old operations command code
