+++ tests/util/test_cli.py
@@ -21,9 +21,7 @@ from dffml.util.cli.cmd import JSONEncoder, CMD, Parser
 from dffml.util.cli.parser import (
     list_action,
     ParseLoggingAction,
-    ParseOutputSpecsAction,
     ParseInputsAction,
-    ParseRemapAction,
 )
 
 from dffml.util.cli.cmds import ListEntrypoint, ModelCMD
@@ -65,20 +63,6 @@ class TestParseActions(unittest.TestCase):
             action(None, namespace, "WARNING")
             mock_method.assert_called_once_with(level=logging.WARNING)
 
-    def test_output_specs(self):
-        namespace = Namespace(output_specs=False)
-        action = ParseOutputSpecsAction(dest="output_specs", option_strings="")
-        action(None, namespace, ["['result']=get_single_spec"])
-        self.assertEqual(len(namespace.output_specs), 1)
-        self.assertEqual(
-            namespace.output_specs[0], (["result"], "get_single_spec")
-        )
-        action(None, namespace, "['result']=get_single_spec")
-        self.assertEqual(len(namespace.output_specs), 1)
-        self.assertEqual(
-            namespace.output_specs[0], (["result"], "get_single_spec")
-        )
-
     def test_inputs(self):
         namespace = Namespace(inputs=False)
         action = ParseInputsAction(dest="inputs", option_strings="")
@@ -88,20 +72,6 @@ class TestParseActions(unittest.TestCase):
         action(None, namespace, "['result']=get_single_spec")
         self.assertEqual(len(namespace.inputs), 1)
         self.assertEqual(namespace.inputs[0], (["result"], "get_single_spec"))
-
-    def test_remap(self):
-        namespace = Namespace(inputs=False)
-        action = ParseRemapAction(dest="inputs", option_strings="")
-        action(None, namespace, ["get_single.result=string_calculator"])
-        self.assertEqual(len(namespace.inputs), 1)
-        self.assertEqual(
-            namespace.inputs[0], ("get_single", "result", "string_calculator")
-        )
-        action(None, namespace, "get_single.result=string_calculator")
-        self.assertEqual(len(namespace.inputs), 1)
-        self.assertEqual(
-            namespace.inputs[0], ("get_single", "result", "string_calculator")
-        )
 
 
 class TestJSONEncoder(unittest.TestCase):
